rename max_pair_distance file to pair_distance since it has both min and max functions in it
